Optimized, and isolated some templates
